fix: #6734 integrations moved on wizard folder
