+++ apps/gauzy/src/app/pages/integrations/integrations-routing.module.ts
@@ -4,26 +4,38 @@ import { IntegrationsComponent } from './integrations.component';
 
 const routes: Routes = [
 	{
-		path: '',
+		path: 'new',
 		component: IntegrationsComponent
 	},
+	/** Integrations Wizard List */
 	{
-		path: 'upwork',
-		loadChildren: () => import('../upwork/upwork.module').then(
-			(m) => m.UpworkModule
-		)
-	},
-	{
-		path: 'hubstaff',
-		loadChildren: () => import('../hubstaff/hubstaff.module').then(
-			(m) => m.HubstaffModule
-		)
-	},
-	{
-		path: 'gauzy-ai',
-		loadChildren: () => import('./gauzy-ai/gauzy-ai.module').then(
-			(m) => m.GauzyAIModule
-		)
+		path: 'wizard',
+		children: [
+			{
+				path: 'upwork',
+				loadChildren: () => import('../upwork/upwork.module').then(
+					(m) => m.UpworkModule
+				)
+			},
+			{
+				path: 'hubstaff',
+				loadChildren: () => import('../hubstaff/hubstaff.module').then(
+					(m) => m.HubstaffModule
+				)
+			},
+			{
+				path: 'gauzy-ai',
+				loadChildren: () => import('./wizard/gauzy-ai/gauzy-ai.module').then(
+					(m) => m.GauzyAIModule
+				)
+			},
+			{
+				path: 'github',
+				loadChildren: () => import('./wizard/github/github.module').then(
+					(m) => m.GithubModule
+				)
+			}
+		]
 	}
 ];
