+++ apps/gauzy/src/app/pages/integrations/integrations.component.html
@@ -88,7 +88,7 @@
 						<a
 							class="list-item"
 							[ngClass]="integration.isComingSoon ? 'disabled': ''"
-							[routerLink]="[integration.redirectUrl]"
+							[routerLink]="['../wizard', integration.redirectUrl]"
 						>
 							<img
 								[src]="integration.fullImgUrl"
