+++ apps/gauzy/src/app/pages/integrations/wizard/gauzy-ai/gauzy-ai-authorize/gauzy-ai-authorize.component.scss
