+++ apps/gauzy/src/app/pages/integrations/wizard/gauzy-ai/gauzy-ai-authorize/gauzy-ai-authorize.component.ts
@@ -4,7 +4,7 @@ import { FormGroup, FormBuilder, Validators, FormGroupDirective } from '@angular
 import { filter, tap } from 'rxjs/operators';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { IIntegrationSetting, IIntegrationTenant, IOrganization, IntegrationEnum } from '@gauzy/contracts';
-import { GauzyAIService, IntegrationsService, Store } from './../../../../@core/services';
+import { GauzyAIService, IntegrationsService, Store } from './../../../../../@core/services';
 
 @UntilDestroy({ checkProperties: true })
 @Component({
