+++ apps/gauzy/src/app/pages/integrations/wizard/gauzy-ai/gauzy-ai.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NbButtonModule, NbCardModule, NbIconModule, NbInputModule, NbTooltipModule } from '@nebular/theme';
-import { TranslateModule } from './../../../@shared/translate/translate.module';
-import { BackNavigationModule } from './../../../@shared/back-navigation';
+import { TranslateModule } from './../../../../@shared/translate/translate.module';
+import { BackNavigationModule } from './../../../../@shared/back-navigation';
 import { GauzyAIRoutingModule } from './gauzy-ai-routing.module';
 import { GauzyAIAuthorizeComponent } from './gauzy-ai-authorize/gauzy-ai-authorize.component';
