+++ packages/core/src/tasks/task.module.ts
@@ -11,7 +11,6 @@ import { TenantModule } from '../tenant/tenant.module';
 import { UserModule } from './../user/user.module';
 import { RoleModule } from './../role/role.module';
 import { EmployeeModule } from './../employee/employee.module';
-import { GitHubModule } from './../github/github.module';
 
 @Module({
 	imports: [
@@ -22,8 +21,7 @@ import { GitHubModule } from './../github/github.module';
 		RoleModule,
 		EmployeeModule,
 		OrganizationProjectModule,
-		CqrsModule,
-		GitHubModule
+		CqrsModule
 	],
 	controllers: [TaskController],
 	providers: [TaskService, ...CommandHandlers],
